Changed stuff to be framerate-independant.

+++ gameStateBased/src/com/cyclight/Projectile.java
@@ -11,9 +11,9 @@ public class Projectile{
 	
 	Shape projShape;
 	boolean projDir;
-	int projSpeed;
+	float projSpeed;
 	
-	public Projectile(Shape shape, boolean direction, int speed)
+	public Projectile(Shape shape, boolean direction, float speed)
 	{
 		projShape=shape;
 		projSpeed=speed;
